Wired it all up, and it lives!
So, this brings the various *_experiment files up to speed feature-wise with
what I already had in example.py, but with substantially more code >_>

Well, no that isn't entirely true.  This version validates the program for
correctness!  And it also has better error messages!  I think this will also
scale up a bit better.

I think the next order of business is a biiiiig cleanup, and then I think maybe
adding support for textures next would be a good idea.