Abstract Syntax Tree Machinery
At this time, codegen is broken again.

This commit revises the grammar rule validation system to automatically
produce a nice AST graph when validation is successful.

This has some handy machinery for providing a nice interface for inspecting
said graph, and should be relatively low friction for updating in the future.

This will hopefully replace what I have in intermediary.py.

I'm planning on adding validaiton checks into the Syntax classes, with the
intention of the result being both a useful intermediary form as well as a
semantically valid program.

Following this would need to be a solver pass and the code gen pass.

Eventually, I imagine that the stuff in graffeine/syntax/ will effectively be
the API-independent frontend, and the solver / expander stuff will be
provided by various API-specific backends.

I don't plan on supporting changing backends at runtime, so the frontend syntax
is only consistent for the sake of easy porting.  Similarly, the lack of any
common shading language means that switching backends is probably going to be
a pain anyway.