Structs can have arrays now, sorta...
Granted, they're broken.  The array in both the generated c++ and glsl structs
always have a size of zero for some reason.  Also the structs with arrays and
structs with structs with arrays and structs with structs with structs with uh,
well, the uploader for those just doesn't upload anything for some reason, even
the non-array and non-struct-with-array etc fields.