WebGL User Vars and Extension Setup
This adds some code to check for required extensions, and copy the provided
functions and symbols into the "gl" global.  This rewrites the names of the
copied properties to remove the extension suffixes.

The following extensions are currently required:

 - OES_vertex_array_object, which provides the ability to create vertex
   array objects.  This is needed by the DefaultVAO expander.

 - ANGLE_instanced_arrays, which provides instanced drawing.  This is needed
   by the InstancedDraw expander.

 - OES_standard_derivatives, because because.

Thinking ahead a bit, there are a number of extensions that look like things
I will definitely want to use, but we have enough information in the renderer
definition to determine in advance if the extension actually needs to be
required.  For example,

 - OES_texture_float and WEBGL_color_buffer_float enable the use of formats
   RGB32F and RGBA32F with textures and render targets respectively.

 - OES_texture_half_float and EXT_color_buffer_half_float enable the use of
   formats RGB16F and RGBA16F with textures and render targets respectively.

There are a number of extensions to provide common filtering modes.  If WebGL
doesn't already have a graceful failure path for the absence of these, it
seems like it could be reasonable to log a warning and fail over to a cheaper
filter:

 - EXT_texture_filter_anisotropic
 - OES_texture_float_linear
 - OES_texture_half_float_linear