WebGL - Shader Compilation and Linking Functions
These are similar to the OpenGL equivalents in dream_machine.hpp, except that
CompileShader's path is the key for fetching the source out of ShaderSources,
LinkShaders currently just takes a list of shader handles to try to link, and
both of them give subpar output at the moment when compilation fails.

This isn't actually hooked up to anything yet, though.

Note for later, `CompileShader("splat.vs", gl.VERTEX_SHADER);` will need to
be stuck somewhere to build the default shader.  Also once the shaders are
compiled, we should probably just delete ShaderSources since there is no need
to keep it around in memory.