WebGL - Shaders and Basic Drawing Capabilities!
The following are now supported in the WebGL backend:

 1. `pipeline` definitions which only declare a fragment shader.

 2. `renderer` definitions which only contain draw events.

 3. `draw` events.

The result is a full screen draw with whatever fragment shader you provided.

A note for future improvement:  Attribute locations can be queried and cached
after a shader is successfully linked.  I'll probably end up doing this when
I add support for uniform buffers.