WebGL - Hacky Uniform Uploading
This adds basic support for uniform uploading.  This only supports plain glsl
types... and possibly more than WebGL might support (eg ivec).  Arrays and
structs are possible with this approach, but I don't really have cause to
bother yet.

WebGL 1 doesn't support uniform buffer objects, which makes implementing
uniforms a bit problematic for how I want Dream Machine to work.  I expect this
will take a few iterations to get right.

The approach I took in this implementation is that for each pipeline that
would be relevant to the uniform upload, the upload function makes that
pipeline's shader the current program, and then calls all of the gl.Upload* etc
functions to upload the data.  This should Just Work, since uniforms are meant
to persist as part of the shader program.  This of course comes at the cost of
requiring to upload the data once per pipeline that uses it.

Also this implementation doesn't actually take account for the dream machine
buffer handles for determining which pipelines are relevant.  This could be a
problem if there were support for uploading buffers with custom data, and two
buffers with the same struct were uploaded with different data.

A final note, uniform block instance names should be pretty easy to make work
here in terms of shader interface, since you just have the struct and one
uniform declaration for that struct.

So, problems for another night:
 1. Maybe generate an uploader per buffer instead of per struct?  And pass in
    the shader handle as an arg, and make the event handler generate the needed
    upload invocations.
 2. Array and struct members aren't implemented yet.
 3. ElapsedTime has the wrong scale.
 4. WindowScale.
 5. Uniform block instance names?  Requires new syntax.
 6. Perf.