WebGL - Determine Required Extensions From Features Used
This change makes some corrections to how formats are propogated so that the
WebGL backend produces the correct strings.

But more importantly, this change makes floating point textures possible with
the WebGL backend, as it automates determining when the extension should be
installed.  So if you use a floating point texture in your program, then the
"OES_texture_float" extension will be required.

This also checks for other extension conditions, eg a floating point texture
with a linear sampler also requires "OES_texture_float_linear".

The spec says "WEBGL_color_buffer_float" and "EXT_color_buffer_half_float" are
"implied" by "OES_texture_float" and "OES_texture_half_float", but my system
makes them explicit as well.  It might be worthwhile in the future to revise
this so that these are only actually needed when there's a render target that
needs them.