Automatic Renderer Switching Syntax
This change adds the "next" command, which indicates that once the containing
renderer finishes running, it should change the current renderer to the one
specified in the "next" command.

For example:

(renderer Setup
	...
	(next Loop))

(renderer Loop
	...)

Whichever renderer is declared first is the starting renderer.  In this case,
"Setup" will be what the application tries to draw first, and then all
subsequent frames will draw the "Loop" renderer.