OpenGL - Pipeline Output Rewrites
This change makes it so pipelines can safely have inputs and outputs which
are declared as sharing a backing texture.

What will happen new texture entries will be added and the relevant outputs
will be rewritten to use those copies.  After a draw uses the pipeline, the
handles for the scratch texture and the original will be switched, and the
framebuffer for the pipeline will be regenerated.

This allows for recursive rendering like so:

(pipeline Accumulate
	(vs "shaders/splat.vs.glsl")
	(fs "shaders/fnord.fs.glsl")
	(in WindowParams)
	(in Accumulator)
	(out Accumulator))

(pipeline Resolve
	(vs "shaders/splat.vs.glsl")
	(fs "shaders/copy.fs.glsl")
	(in WindowParams)
	(in Accumulator))

(renderer Accumulate
	(update WindowParams)
	(draw Accumulate)
	(draw Resolve))

This should be reasonably speedy, but the main problem is there is no mechanism
for recycling the scratch targets between pipeline stages.

I don't really have any specific plan to reconcile this with ARB_texture_barrier
or NV_texture_barrier, but I also don't have any specific plans to add special
support for either extension.

I don't expect this to cause problems for eventually adding support for OpenGL
image objects or UAVs, since I'm planning on using different syntax for that
interface.