Common Enums For Texture And Render Target Formats
This resolves Issue #13.

Format names are now defined in dream_machine/syntax/constants.py, by the enum
class TextureFormats.  This presently is a subset of DXGI_FORMAT, but with a
naming scheme where I can tell the difference between two similar types at a
glance.

The "TYPELESS" variants will not be supported, as I think that is something
dream machine should be able to infer.

Additionally, I left out a bunch of types that aren't immediately needed but
might be desired at a later date.

There is some weirdness with the OpenGL and WebGL types, since the DXGI formats
don't have an equivalent to GL_RG8 or GL_RG8_SNORM and others.

The other oddity is I'm not actually sure what WebGL's gl.DEPTH_COMPONENT is
meant to correspond to, so I arbitrarily picked GL_DEPTH_COMPONENT16 to be the
equivalent.  I'm guessing it's just the 24 bit depth / 8 bit stencil format but
with the stencil ignored.  If that is the case, "DXGI_FORMAT_D24_UNORM_S8_UINT"
would more accurately map to gl.DEPTH_STENCIL and there is no other D24 format
to map to gl.DEPTH_COMPONENT.  But, I'm avoiding supporting combined
depth/stencil formats at the moment, so, whatever.

If need be, I can later add negative numbered formats for weird stuff I guess.