Moved Shadow Texture Creation and Rewrite Helpers Out From OpenGL Backend
This change moves code I'm expecting to copy a bunch of places out of the
OpenGL backend and into the compiler frontend.

This is for issue #4, but will benefit the planned Vulkan and D3D12 backends.

Backends will still have to handle output aliasing in the generated shader code
and flipping texture handles for pipeline outputs w/ shadow targets.