Added Frontend Solve Stage For Implicit Render Events
This is in the `Renderer` syntax class, in the `solve_implicit_steps` function.
So far this just adds the new `RendererTextureSwap` and
`RendererRegenFrameBuffer` events used when a render target needs to be double
buffered.

In the near term, this is intended to slightly reduce the redundant decisioning
between the OpenGL and WebGL backends.  See issue #5.

This will likely be expanded upon in the future for solving things like
resource barriers and heap residency events etc for the explicit API backends.