WebGL - Texture Clearing
Followup to 798d3f696d563be2acc394a85ce49b45d0af7a30.

This change updates the WebGL backend to support the new texture clearing
syntax.

Since WebGL doesn't have an equivalent to glClearTexImage, this is accomplished
by filling out an array for the texture data and passing that in to
gl.texImage2D.

This is probably acceptable perf-wise, since it only has to happen when the
texture is first created and when the canvas is resized.

This resolves issue #2.