Better User Data Uploading for Uniform Buffer Objects
With this change, each named buffer in the program has an entry in the
"Upload" namespace.  The old uploader functions now exist in the
"UploadAction" namespace (which is not exposed outside generated.cpp), and is
used to facilitate the actual data upload.

In a sense, the new upload functions basically wrap the old one, enclosing the
correct buffer handle lookup, and route the user's data along.

Uploading happens as soon as this is called, so, it is probably only really
safe to call by the user within the user frame callback, but this is not
enforced at the moment.

This is part of issue #17.  This should be followup by a change to remove the
RendererUpdate syntax object etc since I have no plans to use it for anything
meaningful at the moment.

I'd also like to follow this up w/ a similar data uploading API for texture data,
which maybe shouldn't be too hard?

Stuff like SSBOs and VBOs will probably be different issue tickets, but I think
this at least helps inform future design decisions for these things.