OpenGL Image Load/Store Semantics
This commit adds support for Image Load/Store semantics to the OpenGL backend.

This is used similarly to the existing texture support, and the same high level
syntax will be used for targets w/ similar concepts (eg, D3D UAVs).

To use this, create a texture as you normally would, but instead of using the
"in" keyword in a pipeline definition, use the new "side" keyword.  The shaders
associated to the pipeline will have the image variables set up automatically.

You currently need to add "#extension GL_ARB_shader_image_load_store : require"
to the top of your shader source manually however.

This commit resolves issue #16.