Removed old implementation.
The goal for this branch will be to produce a cython library library
for prototyping rendering techniques and simple games.

The specifics are up in the air at the moment, and this change is
mainly to provide a clean base to work off of.

I'm planning on using SDL2 for window management and input events,
and probably just Vulkan as the rendering backend.  Other backends
might make sense in the future, provided compatible semantics.

My primary goals here are now:

 1. Easy high level pythony API for resource management and scheduling
    rendering work.

 2. Automatic barriers etc.

 3. Fast enough to do useful profiling of rendering prototypes.

 4. Expose modern rendering features, like ray tracing, mesh shaders,
    and variable rate shading.

 5. Live programming for graphics!

 6. Reasonable compatibility with python libraries like pillow and
    numpy.

The rough design I have is one builds a rendering pipeline out of a
list of event objects.  Event objects are for the most part going
to be various kinds of draw / dispatch events, and the appropriate
resource bindings.

When sequences of events are registered, the program is returned a
handle for the entire pipeline, which is to be called when it is
desired to draw a new frame.

With this, a basic program might look something like this:

 1. SDL2 / Vulkan initialization

 2. Create GPU resource buffers.

 3. Upload, say, textures etc to said buffers.

 4. Create event objects, fill in parameters.

 5. Register an event list, get a handle.

 6. Execute the handle over and over again.

This design allows for adjustments to be made on the fly via uploading
new uniforms etc, without having to regenerate event lists.

It makes sense to support registering multiple event lists to make
switching between different program states relatively easy, or for
utility purposes like mip generation.

While I want registration to be reasonably quick, a program's steady
state should generally avoid registering new event lists in favor of
using things like multi-draw/dispatch parameters and uniform vars to
control rendering.

Event object resource bindings are likely to be introspected from
shaders.  Etc.